+++ src/components/manager/main.js
@@ -2,3 +2,4 @@ export * from './ChartManager';
 export { createChart } from './ChartManager/utils';
 export * from './Responsive';
 export * from './TableManager';
+export * from './ElasticWidget';